String calculation attempt #2 (still flawed)

+++ calculate.hpp
@@ -7,4 +7,4 @@
 
 #pragma once
 
-int calculate(const std::string& mathString);
+float calculate(const std::string& mathString);
